The directory structure is changed, OnePathogenSimu is refactored, and test classes are added.
