+++ pom.xml
@@ -24,5 +24,11 @@
             <artifactId>jfreechart</artifactId>
             <version>1.5.1</version>
         </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.12</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 </project>
