+++ src/main/java/Business/DrawArea/Area.java
@@ -6,9 +6,9 @@ public class Area {
     public Area(){}
 
     public Area(int areaLength, int areaWidth){
-        this.area = new AreaUnit[areaWidth/AreaUnit.BLOCK_WIDTH][areaLength/AreaUnit.BLOCK_LENGTH];
-        for(int i=0;i<areaWidth/AreaUnit.BLOCK_WIDTH;i++){
-            for(int j=0;j<areaLength/AreaUnit.BLOCK_LENGTH;j++){
+        this.area = new AreaUnit[areaWidth/ AreaUnit.BLOCK_WIDTH][areaLength/ AreaUnit.BLOCK_LENGTH];
+        for(int i = 0; i<areaWidth/ AreaUnit.BLOCK_WIDTH; i++){
+            for(int j = 0; j<areaLength/ AreaUnit.BLOCK_LENGTH; j++){
                 area[i][j] = new AreaUnit();
             }
         }
