+++ src/main/java/Business/DrawArea/AreaUnit.java
@@ -20,7 +20,12 @@ public class AreaUnit {
     }
 
     public void calcInfectSpeed(){
-        this.infectSpeed = 1.0;
+        calcPopulationDensity();
+        this.infectSpeed = infectNum * (0.005 * populationDensity * pathogen.getR_FACTOR() * 0.3 / pathogen.getK_FACTOR() * 9);
+        //this.infectSpeed = 1.0;
+    }
+    private void calcPopulationDensity(){
+        this.populationDensity = this.headcount / (BLOCK_WIDTH * BLOCK_LENGTH);
     }
 
     public double getPopulationDensity() {
