+++ src/main/java/Business/Pathogen/Hospital.java
