+++ src/main/java/Business/Plot/LineChart.java
@@ -1,0 +1,55 @@
+package Business.Plot;
+
+import org.jfree.chart.ChartFactory;
+import org.jfree.chart.ChartUtils;
+import org.jfree.chart.JFreeChart;
+import org.jfree.chart.plot.PlotOrientation;
+import org.jfree.chart.title.TextTitle;
+import org.jfree.data.xy.XYSeries;
+import org.jfree.data.xy.XYSeriesCollection;
+
+import java.awt.*;
+import java.io.FileOutputStream;
+
+public class LineChart {
+
+    public static void main(String[] args) {
+
+        XYSeries series = new XYSeries("xySeries");
+        for (int x = -100; x < 100; x++) {
+            int y = x * x;
+            series.add(x, y);
+        }
+
+        XYSeriesCollection dataset = new XYSeriesCollection();
+        dataset.addSeries(series);
+        JFreeChart chart = ChartFactory.createXYLineChart(
+                "y = x^2", // chart title
+                "x", // x axis label
+                "x^2", // y axis label
+                dataset, // data
+                PlotOrientation.VERTICAL,
+                false, // include legend
+                false, // tooltips
+                false // urls
+        );
+
+        Font font = new Font("Arial",Font.CENTER_BASELINE,20);
+        TextTitle title = new TextTitle("test");
+        title.setFont(font);
+        chart.setTitle(title);
+        FileOutputStream fos_jpg = null;
+        try {
+            fos_jpg = new FileOutputStream("test.jpg");
+            ChartUtils.writeChartAsJPEG(fos_jpg,0.7f,chart,640,480,null);
+            fos_jpg.close();
+        } catch (Exception e) {
+        }
+
+//        ChartFrame frame = new ChartFrame("my picture", chart);
+//        frame.pack();
+//        frame.setVisible(true);
+//        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+
+    }
+}
