+++ src/main/java/UserInterface/ViewReport/ChooseSimuJPanel.java
